Rearrange Monoid
